# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "beego Test API",
        "description": "beego has a very cool tools to autogenerate documents for your API",
        "version": "1.0.0",
        "termsOfService": "http://beego.me/",
        "contact": {
            "email": "astaxie@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/lastest",
    "paths": {
...
        "/webapi//createqrcode": {
            "post": {
                "tags": [
                    "webapi"
                ],
                "description": "create users",
                "operationId": "WechatController.CreateArticle",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "body for user content",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Article"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{int} models.Article.Id"
                    },
                    "403": {
                        "description": "body is empty"
                    }
                }
            }
        },
...
        "/wechatapi/small/article/": {
            "get": {
                "tags": [
                    "wechatapi/small/article"
                ],
                "description": "get all Articles",
                "operationId": "ArticleController.GetAll",
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.Article"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "wechatapi/small/article"
                ],
                "description": "create users",
                "operationId": "ArticleController.CreateArticle",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "body for user content",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Article"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{int} models.Article.Id"
                    },
                    "403": {
                        "description": "body is empty"
                    }
                }
            }
        },
...
        "/wechatapi/small/config/": {
            "get": {
                "tags": [
                    "wechatapi/small/config"
                ],
                "description": "get user by uid",
                "operationId": "ConfigController.Get",
                "parameters": [
                    {
                        "in": "path",
                        "name": "uid",
                        "description": "The key for staticblock",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.Article"
                        }
                    },
                    "403": {
                        "description": ":uid is empty"
                    }
                }
            },
            "post": {
                "tags": [
                    "wechatapi/small/config"
                ],
                "description": "create users",
                "operationId": "ConfigController.CreateArticle",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "body for user content",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Article"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{int} models.Article.Id"
                    },
                    "403": {
                        "description": "body is empty"
                    }
                }
            }
        },
...
        "/wechatapi/small/location/": {
            "post": {
                "tags": [
                    "wechatapi/small/location"
                ],
                "description": "create users",
                "operationId": "LocationController.CreateArticle",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "body for user content",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Article"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{int} models.Article.Id"
                    },
                    "403": {
                        "description": "body is empty"
                    }
                }
            }
        },
...
        "/wechatapi/wechat//createqrcode": {
            "post": {
                "tags": [
                    "webapi"
                ],
                "description": "create users",
                "operationId": "WechatController.CreateArticle",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "body for user content",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Article"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{int} models.Article.Id"
                    },
                    "403": {
                        "description": "body is empty"
                    }
                }
            }
        },
...
        "/wechatapi/wechat/template/message/": {
            "get": {
                "tags": [
                    "wechatapi/wechat/template/message"
                ],
                "description": "get user by uid",
                "operationId": "SendWechatTemplateMessage.Get",
                "parameters": [
                    {
                        "in": "path",
                        "name": "uid",
                        "description": "The key for staticblock",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.Article"
                        }
                    },
                    "403": {
                        "description": ":uid is empty"
                    }
                }
            },
            "post": {
                "tags": [
                    "wechatapi/wechat/template/message"
                ],
                "description": "create users",
                "operationId": "SendWechatTemplateMessage.CreateArticle",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "body for user content",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Article"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{int} models.Article.Id"
                    },
                    "403": {
                        "description": "body is empty"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Article": {
            "title": "Article",
            "type": "object",
            "properties": {
                "Address": {
                    "type": "string"
                },
                "Age": {
                    "type": "integer",
                    "format": "int64"
                },
                "Arcid": {
                    "type": "string"
                },
                "AvatarUrl": {
                    "type": "string"
                },
                "Babyid": {
                    "type": "integer",
                    "format": "int64"
                },
                "BirthedAddress": {
                    "type": "string"
                },
                "BirthedAt": {
                    "type": "string",
                    "format": "datetime"
                },
                "BirthedCity": {
                    "type": "string"
                },
                "BirthedCountry": {
                    "type": "string"
                },
                "BirthedProvince": {
                    "type": "string"
                },
                "Category": {
                    "type": "string"
                },
                "Characters": {
                    "type": "string"
                },
                "City": {
                    "type": "string"
                },
                "Country": {
                    "type": "string"
                },
                "DataFrom": {
                    "type": "string"
                },
                "Details": {
                    "type": "string"
                },
                "Gender": {
                    "type": "integer",
                    "format": "int32"
                },
                "Handler": {
                    "type": "string"
                },
                "Height": {
                    "type": "string"
                },
                "MissedAddress": {
                    "type": "string"
                },
                "MissedAt": {
                    "type": "string",
                    "format": "datetime"
                },
                "MissedCity": {
                    "type": "string"
                },
                "MissedCountry": {
                    "type": "string"
                },
                "MissedProvince": {
                    "type": "string"
                },
                "Nickname": {
                    "type": "string"
                },
                "Province": {
                    "type": "string"
                },
                "Title": {
                    "type": "string"
                },
                "UUID": {
                    "type": "string"
                }
            }
        },
...
    },
...
}